fix: double -> BigDecimal 타입 변경

+++ m-common/src/main/java/com/antmen/antwork/common/api/response/ManagerReviewInfoDto.java
@@ -16,7 +16,7 @@ public class ManagerReviewInfoDto {
     private Integer totalWrittenReviews;
     private Integer totalReceivedReviews;
     private BigDecimal averageWrittenRating;
-    private Double averageReceivedRating;
+    private BigDecimal averageReceivedRating;
     private List<CustomerReviewDto> writtenReviews;
     private List<CustomerReviewDto> receivedReviews;
 }
